fix: Simplified selecting chat
Previously I was using useState to select chat but it was unnecessary as it can we achieved using useRouter much easily. This way I dont need to pass selectedChat & setSelectedChat as props to chat component, we can just set chatId query parameter in the url then use useRouter to see which chatId is currently selected easier & better than before.